Add main script
The program analyses the places.sqlite database that firefox uses to store
bookmarks and history data.

For google links, the programs removes the tracking data contained in the
urls, leaving only the domain and the search query. For youtube links, the
program retrieves the channel name and the video's title.

There are a couple of hard coded values. The timestamp indicates the date:
2020-04-01 00:00.

Signed-off-by: Victor Otieno <vickz84259@gmail.com>